GCP_cloud_functions/rsu-manager/rsu-querycounts/requirements.txt
DateTime==4.3
google-cloud-bigquery==2.23.2